Remove only duplicated data centers without validators or gossip nodes
